added CityDto added insert and find methods into city-controller

+++ src/main/resources/db/migration/V001__create_table_city.sql
@@ -3,5 +3,5 @@ CREATE TABLE city
     id                 INTEGER PRIMARY KEY,
     name               VARCHAR(300) NOT NULL,
     date_of_foundation DATE,
-    population         INTEGER
+    population         BIGINT
 );
